Rename charmstore directory to repositories, used by more than charmstore code.
